made the files more modular and completed the About Us Page
