+++ package.json
@@ -7,7 +7,7 @@
     "firebase": "^9.6.10",
     "react": "^17.0.1",
     "react-dom": "^17.0.1",
-    "react-router-dom": "^6.3.0",
+    "react-router-dom": "^5.3.3",
     "react-scripts": "4.0.1"
   },
   "devDependencies": {
