+++ src/index.js
@@ -1,13 +1,12 @@
-// import React from 'react';
 import ReactDOM from 'react-dom';
 import { BrowserRouter } from 'react-router-dom';
 import './index.css';
-import App from './App';
+import App2 from './App2';
 import * as serviceWorker from './serviceWorker';
 
 ReactDOM.render(
   <BrowserRouter>
-    <App />
+    <App2 />
   </BrowserRouter>,
   document.getElementById('root'),
 );
