+++ src/pages/Firebase/firebase.js
@@ -1,4 +1,4 @@
-import { db } from '../lib/firebase';
+import { db } from '../../lib/firebase';
 import { collection, getDocs, addDoc } from 'firebase/firestore';
 import { useEffect, useState } from 'react';
